Set the deployment target to 10.9.
Poke, prod, improve:
Move the iVars for TCMPortMapper to its implementation block.
Minor modernization.
Fix IXSCNotificationManager's improper migration to ARC: The use of CFBridgingRelease in _IXSCNotificationCallback would have caused big problems.

+++ framework/libnatpmp/natpmp.c
@@ -168,7 +168,7 @@ int readnatpmpresponse(natpmp_t * p, natpmpresp_t * response)
 	int n;
 	if(!p)
 		return NATPMP_ERR_INVALIDARGS;
-	n = recvfrom(p->s, buf, sizeof(buf), 0,
+	n = (int)recvfrom(p->s, buf, sizeof(buf), 0,
 	             (struct sockaddr *)&addr, &addrlen);
 	if(n<0)
 		switch(errno) {
